bot/requirements.txt
python-telegram-bot>=20.0,<21.0
requests>=2.28.0,<3.0.0
httpx>=0.25.0,<0.26.0